# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Management API
  description: API REST pour la gestion de plantes domestiques
  version: 1.0.0
  contact:
    name: Plant Team
servers:
  - url: http://localhost:8080/api/v1
    description: Development (Quarkus local)
  # a remplacer par l'URL du cluster Kubernetes université ?
  - url: http://plant-backend.etu-k8s.lan/api/v1
    description: Production (Kubernetes cluster)
...
security:
  - bearerAuth: []
paths:
...
  /rooms:
    get:
      tags: [Room]
      summary: Lister les pièces de ma maison
      responses:
        '200':
          description: Liste des pièces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'
    post:
      tags: [Room]
      summary: Créer une pièce
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: Pièce créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
  /rooms/{id}:
    get:
      tags: [Room]
      summary: Détails d'une pièce
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Détails de la pièce
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetailResponse'
    patch:
      tags: [Room]
      summary: Modifier une pièce
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: Pièce modifiée
    delete:
      tags: [Room]
      summary: Supprimer une pièce
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Pièce supprimée

  # ==================== SPECIES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mcpApiKey:
      type: apiKey
      in: header
      name: X-MCP-API-Key
      description: API Key pour l'accès MCP (Goose/Mistral)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateRoomRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        type:
          type: string
          enum: [LIVING_ROOM, BEDROOM, BALCONY, GARDEN, KITCHEN, BATHROOM, OFFICE, OTHER]
    UpdateRoomRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [LIVING_ROOM, BEDROOM, BALCONY, GARDEN, KITCHEN, BATHROOM, OFFICE, OTHER]
    RoomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        plantCount:
          type: integer
    RoomDetailResponse:
      allOf:
        - $ref: '#/components/schemas/RoomResponse'
        - type: object
          properties:
            plants:
              type: array
              items:
                $ref: '#/components/schemas/PlantResponse'

    # ===== SPECIES =====
    SpeciesResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trefleId:
          type: integer
        commonName:
          type: string
        scientificName:
          type: string
        family:
          type: string
        imageUrl:
          type: string
...
    PlantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        species:
          $ref: '#/components/schemas/SpeciesResponse'
        room:
          $ref: '#/components/schemas/RoomResponse'
        photoUrl:
          type: string
        lastWatered:
          type: string
          format: date-time
        nextWatering:
          type: string
          format: date-time
        needsWater:
          type: boolean
        acquiredAt:
          type: string
          format: date
...